+++ command/statistics/index.js
@@ -13,10 +13,13 @@ const MODULE_REQUIRE = 1
 
 	, getDownloadCount = require('npm-registry-rest/getDownloadCount')
 	, getPackageNames = require('npm-registry-rest/getPackageNames')
+	, getPackage = require('npm-registry-rest/getPackage')
+	, getStarCount = require('npm-registry-rest/getStarCount')
 
 	/* in-jinang */
 	, co = require('jinang/co')
 	, table = require('jinang/table')
+	, sort = require('jinang/sort')
 	
 	/* in-package */
 	, config = noda.inRequire('lib/config')
@@ -27,7 +30,16 @@ function help() {
 }
 
 function package_stat(name) { co(function*() {
-	console.log('name:', colors.bold(name));
+	console.log('package name:', colors.bold(name));
+
+	let meta = yield getPackage(name);
+	console.log('-- stars --');
+	if (meta.users) {
+		console.log(Object.keys(meta.users).length);
+	}
+	else {
+		console.log('No users starred this package.');
+	}
 
 	console.log('-- downloads --');
 	let ranges = [ 'last-day', 'last-week', 'last-month'];
@@ -37,9 +49,11 @@ function package_stat(name) { co(function*() {
 		console.log(range, '\t', colors.bold(count));
 	}
 
+}).catch(err => {
+	console.error(`failed to find package: ${name}`);
 }); }
 
-function owner_stat(username) { co(function*() {
+function owner_stat(username, orderby) { co(function*() {
 	if (username[0] == '~') username = username.slice(1);
 
 	console.log('username:', colors.bold(username));
@@ -49,6 +63,19 @@ function owner_stat(username) { co(function*() {
 	process.stdout.clearLine();
 	process.stdout.cursorTo(0);
 	console.log('total packages:', colors.bold(packageNames.length));
+
+	if (orderby == 'name') {
+		packageNames.sort();
+	}
+
+	// let stars = {};
+	// for (let i = 0; i < packageNames.length; i++) {
+	// 	let name = packageNames[i];
+	// 	process.stdout.write(`querying package stars: ${name}`);
+	// 	stars[name] = yield getStarCount(name);
+	// 	process.stdout.clearLine();
+	// 	process.stdout.cursorTo(0);
+	// }
 	
 	const range = 'last-day';
 	const counts = {};
@@ -62,6 +89,8 @@ function owner_stat(username) { co(function*() {
 	}
 
 	let rows = [];
+	let total = { name: 'TOTAL' };
+	ranges.forEach(range => total[range] = 0);
 	packageNames.forEach(name => {
 		let count = counts['last-day'][name];
 		let color;
@@ -85,31 +114,51 @@ function owner_stat(username) { co(function*() {
 		}
 
 		rows.push({
-			name: color(name),
-			'last-day': color(count),
-			'last-week': colors.dim(counts['last-week'][name]),
-			'last-month': colors.dim(counts['last-month'][name]),
+			name           : color(name),
+			// stars          : stars[name],
+			orderDownloads : count,
+			'last-day'     : color(count),
+			'last-week'    : colors.dim(counts['last-week'][name]),
+			'last-month'   : colors.dim(counts['last-month'][name]),
 		});
+
+		ranges.forEach(range => total[range] += counts[range][name]);
 	});
+
+	if (orderby == 'downloads') {
+		sort(rows, 'orderDownloads');
+	}
+
+	rows.push(total);
 
 	console.log('-- downloads --');
 	table.print(rows, {
-		columns: [ 'name', 'last-day', 'last-week', 'last-month' ],
+		columns: [ 
+			'name',
+			{ name: 'last-day', align: 'right' }, 
+			{ name: 'last-week', align: 'right' }, 
+			{ name: 'last-month', align: 'right' },
+			// { name: 'stars', align: 'right' },
+		],
 	});
 }).catch(ex => console.log(ex)); }
 
 function run(argv) {
 	const groups = [ 
 		[ 
-			'--help -h [*:=*help] REQUIRED', 
+			'--help -h [*:=* help] REQUIRED', 
 		], [
-			'--package NOT NULL',
-			'--owner NOT NULL',
-		]
+			'--package NOT NULL REQUIRED',
+		], [
+			'--owner NOT NULL REQUIRED',
+			'--orderby NOT NULL',
+		], [
+			'--name [0] NOT NULL REQUIRED',
+			'--orderby NOT NULL',
+		],
 	];
 	
 	const cmd = commandos.parse.onlyArgs(argv, { groups, catcher: help });
-	const name = if2.defined(cmd.$[0], '');
 
 	if (!cmd) {
 		return;
@@ -121,16 +170,13 @@ function run(argv) {
 		return package_stat(cmd.package);
 	}
 	else if (cmd.owner) {
-		return owner_stat(cmd.owner);
+		return owner_stat(cmd.owner, cmd.orderby);
 	}
-	else if (name[0] == '~') {
-		return owner_stat(name);
+	else if (cmd.name[0] == '~') {
+		return owner_stat(cmd.name, cmd.orderby);
 	}
-	else if (name) {
-		return package_stat(name);
-	}
-	else {
-		return help();
+	else if (cmd.name) {
+		return package_stat(cmd.name);
 	}
 }