+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning 2.0.0](http://semver.org/).
 
+##  [0.1.1] - Mar 29th, 2018
+
+*   Option `--orderby` accepted by sub command `nopa statistics`.
+*   `nopa statistics` will display number of stars on displaying package stat info.
+*   Dependencies upgraded.
+
 ##  [0.1.0] - Mar 26th, 2018
 
 *   Sub command `nopa profile` added.